+++ client/.firebaserc
@@ -1,0 +1,5 @@
+{
+  "projects": {
+    "default": "step9-2020-capstone"
+  }
+}
