+++ client/package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "algoliasearch": "3.35.1",
     "express": "^4.17.1",
+    "firebase": "^7.16.0",
     "path": "^0.12.7",
     "react": "16.12.0",
     "react-dom": "16.12.0",
